[ty] Share the importer with diagnostic autofixes
[ty] Regenerate rule documentation
