+++ crates/ty_ide/Cargo.toml
@@ -20,8 +20,6 @@ ruff_diagnostics = { workspace = true }
 ruff_index = { workspace = true }
 ruff_memory_usage = { workspace = true }
 ruff_python_ast = { workspace = true }
-ruff_python_codegen = { workspace = true }
-ruff_python_importer = { workspace = true }
 ruff_python_literal = { workspace = true }
 ruff_python_stdlib = { workspace = true }
 ruff_python_trivia = { workspace = true }
