+++ crates/ty_python_core/src/ast_node_ref.rs
@@ -71,7 +71,7 @@ where
     ///
     /// This method may panic or produce unspecified results if the provided module is from a
     /// different file, Python version, or Salsa revision than the module to which the node belongs.
-    pub(super) fn new(module_ref: &ParsedModuleRef, node: &T) -> Self {
+    pub fn new(module_ref: &ParsedModuleRef, node: &T) -> Self {
         let index = node.node_index().load();
         debug_assert_eq!(module_ref.get_by_index(index).try_into().ok(), Some(node));
