+++ crates/ty_python_semantic/Cargo.toml
@@ -18,6 +18,8 @@ ruff_macros = { workspace = true }
 ruff_memory_usage = { workspace = true }
 ruff_python_ast = { workspace = true, features = ["salsa"] }
 ruff_python_edits = { workspace = true }
+ruff_python_codegen = { workspace = true }
+ruff_python_importer = { workspace = true }
 ruff_python_literal = { workspace = true }
 ruff_python_parser = { workspace = true }
 ruff_python_stdlib = { workspace = true }
