+++ crates/ty_python_semantic/src/lib.rs
@@ -19,8 +19,7 @@ use ruff_db::parsed::parsed_module;
 use ruff_db::source::{SourceTextError, source_text};
 use rustc_hash::FxHasher;
 pub use semantic_model::{
-    Completion, ExpectedStringLiteralCompletion, HasDefinition, HasType, MemberDefinition,
-    NameKind, SemanticModel,
+    Completion, ExpectedStringLiteralCompletion, HasDefinition, HasType, NameKind, SemanticModel,
 };
 use std::hash::BuildHasherDefault;
 pub use suppression::suppress_single;
@@ -57,6 +56,7 @@ mod db;
 pub mod dependency;
 mod dunder_all;
 mod fixes;
+pub mod importer;
 mod lexical_name_path;
 pub mod lint;
 pub(crate) mod place;
